config: Split config module into several submodules
The reason for the change is that it will become easier to import
error-related functions from buildbot.config.errors without introducing
a circular import.
