+++ master/buildbot/config/__init__.py
@@ -1,0 +1,25 @@
+# This file is part of Buildbot.  Buildbot is free software: you can
+# redistribute it and/or modify it under the terms of the GNU General Public
+# License as published by the Free Software Foundation, version 2.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+# details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, write to the Free Software Foundation, Inc., 51
+# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+#
+# Copyright Buildbot Team Members
+
+
+from .builder import BuilderConfig  # noqa pylint: disable=unused-import
+from .errors import ConfigErrors  # noqa pylint: disable=unused-import
+from .errors import capture_config_errors  # noqa pylint: disable=unused-import
+from .errors import error  # noqa pylint: disable=unused-import
+from .master import FileLoader  # noqa pylint: disable=unused-import
+from .master import MasterConfig  # noqa pylint: disable=unused-import
+from .master import get_is_in_unit_tests  # noqa pylint: disable=unused-import
+from .master import loadConfigDict  # noqa pylint: disable=unused-import
+from .master import set_is_in_unit_tests  # noqa pylint: disable=unused-import
